// src/generated/quiz-service/api.ts
export interface QuestionDTO {
    /**
     * 
     * @type {string}
     * @memberof QuestionDTO
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof QuestionDTO
     */
    'content': string;
    /**
     * 
     * @type {string}
     * @memberof QuestionDTO
     */
    'answer1'?: string;
    /**
     * 
     * @type {string}
     * @memberof QuestionDTO
     */
    'answer2'?: string;
    /**
     * 
     * @type {string}
     * @memberof QuestionDTO
     */
    'answer3'?: string;
    /**
     * 
     * @type {string}
     * @memberof QuestionDTO
     */
    'answer4'?: string;
    /**
     * 
     * @type {string}
     * @memberof QuestionDTO
     */
    'answer5'?: string;
    /**
     * 
     * @type {number}
     * @memberof QuestionDTO
     */
    'correctAnswer'?: number;
    /**
     * 
     * @type {string}
     * @memberof QuestionDTO
     */
    'source'?: string;
    /**
     * 
     * @type {number}
     * @memberof QuestionDTO
     */
    'sourcePage'?: number;
    /**
     * 
     * @type {number}
     * @memberof QuestionDTO
     */
    'difficultyLevel'?: number;
    /**
     * 
     * @type {string}
     * @memberof QuestionDTO
     */
    'hint'?: string;
    /**
     * 
     * @type {string}
     * @memberof QuestionDTO
     */
    'createdBy'?: string;
}